Rename ex-1 -> ex1-volume, ex2 -> ex2-surface and add ex2-surface.okl
