+++ examples/ceed/ex1-volume.okl
@@ -18,34 +18,25 @@
 typedef int CeedInt;
 typedef double CeedScalar;
 
-// OCCA parser doesn't like __global here
-typedef /*__global*/ double gCeedScalar;
-
-// OCCA parser doesn't like struct here
-//typedef /*__global*/ struct BuildContext gBuildContext;
-
-// *****************************************************************************
-// OCCA parser doesn't like struct here either
-//struct BuildContext { CeedInt dim, space_dim; };
-
 // *****************************************************************************
 @kernel void f_build_mass(int *ctx, CeedInt Q,
                           const int *iOf7, const int *oOf7, 
-                          const CeedScalar *u, CeedScalar *v) {
+                          const CeedScalar *in, CeedScalar *out) {
+//  const CeedScalar
+//    *J = in + iOf7[0];
+//    *w = in + iOf7[1];
+//  CeedScalar
+//    *qdata = out + oOf7[0];
   for (int i=0; i<Q; i++; @tile(TILE_SIZE,@outer,@inner)) {
-    // OCCA parser can't insert an __global here
-    /*const CeedScalar *J = u + iOf7[0];
-    const CeedScalar *qw = u + iOf7[1];
-    CeedScalar *qd = v + oOf7[0];*/
     //const struct BuildContext *bc = (struct BuildContext*)ctx;
     const int dim = ctx[0];
     const int space_dim = ctx[1];
     const CeedInt dims = dim + 10*space_dim;
-    if (dims==11) v[oOf7[0]+i] = u[iOf7[0]+i] * u[iOf7[1]+i];
-    if (dims==22) v[oOf7[0]+i] = (u[iOf7[0]+i+Q*0]*u[iOf7[0]+i+Q*3]-u[iOf7[0]+i+Q*1]*u[iOf7[0]+i+Q*2])*u[iOf7[1]+i];
-    if (dims==33) v[oOf7[0]+i] = (u[iOf7[0]+i+Q*0]*(u[iOf7[0]+i+Q*4]*u[iOf7[0]+i+Q*8]-u[iOf7[0]+i+Q*5]*u[iOf7[0]+i+Q*7]) -
-                                  u[iOf7[0]+i+Q*1]*(u[iOf7[0]+i+Q*3]*u[iOf7[0]+i+Q*8]-u[iOf7[0]+i+Q*5]*u[iOf7[0]+i+Q*6]) +
-                                  u[iOf7[0]+i+Q*2]*(u[iOf7[0]+i+Q*3]*u[iOf7[0]+i+Q*7]-u[iOf7[0]+i+Q*4]*u[iOf7[0]+i+Q*6]))*u[iOf7[1]+i];
+    if (dims==11) out[oOf7[0]+i] = in[iOf7[0]+i] * in[iOf7[1]+i];
+    if (dims==22) out[oOf7[0]+i] = (in[iOf7[0]+i+Q*0]*in[iOf7[0]+i+Q*3]-in[iOf7[0]+i+Q*1]*in[iOf7[0]+i+Q*2])*in[iOf7[1]+i];
+    if (dims==33) out[oOf7[0]+i] = (in[iOf7[0]+i+Q*0]*(in[iOf7[0]+i+Q*4]*in[iOf7[0]+i+Q*8]-in[iOf7[0]+i+Q*5]*in[iOf7[0]+i+Q*7]) -
+                                    in[iOf7[0]+i+Q*1]*(in[iOf7[0]+i+Q*3]*in[iOf7[0]+i+Q*8]-in[iOf7[0]+i+Q*5]*in[iOf7[0]+i+Q*6]) +
+                                    in[iOf7[0]+i+Q*2]*(in[iOf7[0]+i+Q*3]*in[iOf7[0]+i+Q*7]-in[iOf7[0]+i+Q*4]*in[iOf7[0]+i+Q*6]))*in[iOf7[1]+i];
   }
 }
 
@@ -53,12 +44,11 @@ typedef /*__global*/ double gCeedScalar;
 @kernel void f_apply_mass(void *ctx, CeedInt Q,
                           const int *iOf7, const int *oOf7,
                           const CeedScalar *in, CeedScalar *out) {
+//  const CeedScalar
+//    *u = in + iOf7[0],
+//    *qdata = in + iOf7[1],
+//  CeedScalar *v = out + oOf7[0];
   for (int i=0; i<Q; i++; @tile(TILE_SIZE,@outer,@inner)) {
-    // OCCA parser can't insert an __global here
-    /*const CeedScalar *u = in + iOf7[0];
-    const CeedScalar *w = in + iOf7[1];
-    CeedScalar *v = out + oOf7[0];
-    v[i] = w[i] * u[i];*/
     out[oOf7[0]+i] = in[iOf7[1]+i] * in[iOf7[0]+i];
   }
 }
